Require a '=' between function name and body
With the unification of function_body and body single expression
functions no longer require that their body be wrapped in {}, i.e.

fun add(a, b) a + b

would be a valid function definition.  This could be confusing, and
also just looks weird.  With this change the definition becomes:

fun add(a, b) = a + b

or equivalently:

fun add(a, b) = {
	a + b;
}